+++ packages/components/src/Combobox/components/ComboboxOption/ComboboxOption.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "option": string;
+};
+export = styles;
+