+++ packages/components/src/Combobox/components/ComboboxTrigger/ComboboxTriggerChip/ComboboxTriggerChip.css.d.ts
@@ -1,6 +1,0 @@
-declare const styles: {
-  readonly "trigger": string;
-  readonly "triggerIcon": string;
-};
-export = styles;
-